+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunsFeedTabs.tsx
@@ -6,7 +6,7 @@ import styled, {css} from 'styled-components';
 
 import {failedStatuses, inProgressStatuses, queuedStatuses} from './RunStatuses';
 import {runsPathWithFilters, useQueryPersistedRunFilters} from './RunsFilterInput';
-import {RunTabsCountQuery, RunTabsCountQueryVariables} from './types/RunListTabs.types';
+import {RunFeedTabsCountQuery, RunFeedTabsCountQueryVariables} from './types/RunsFeedTabs.types';
 import {gql, useQuery} from '../apollo-client';
 import {RunStatus, RunsFilter} from '../graphql/types';
 import {useDocumentTitle} from '../hooks/useDocumentTitle';
@@ -31,8 +31,8 @@ const getDocumentTitle = (selected: ReturnType<typeof useSelectedRunsFeedTab>) =
 };
 
 export const useRunsFeedTabs = (filter: RunsFilter = {}) => {
-  const queryResult = useQuery<RunTabsCountQuery, RunTabsCountQueryVariables>(
-    RUN_TABS_COUNT_QUERY,
+  const queryResult = useQuery<RunFeedTabsCountQuery, RunFeedTabsCountQueryVariables>(
+    RUN_FEED_TABS_COUNT_QUERY,
     {
       variables: {
         queuedFilter: {...filter, statuses: Array.from(queuedStatuses)},
@@ -126,8 +126,8 @@ export const useSelectedRunsFeedTab = (filterTokens: TokenizingFieldValue[]) => 
   return 'all';
 };
 
-export const RUN_TABS_COUNT_QUERY = gql`
-  query RunTabsCountQuery($queuedFilter: RunsFilter!, $inProgressFilter: RunsFilter!) {
+export const RUN_FEED_TABS_COUNT_QUERY = gql`
+  query RunFeedTabsCountQuery($queuedFilter: RunsFilter!, $inProgressFilter: RunsFilter!) {
     queuedCount: pipelineRunsOrError(filter: $queuedFilter) {
       ... on Runs {
         count
